requirements.txt
asgiref==3.3.4
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cloudinary==1.26.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.2
django-appconf==1.0.5
django-configurations==2.2
django-cryptography==1.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-rest-passwordreset==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
facebook-sdk==3.1.0
google-api-core==2.0.1
google-api-python-client==2.18.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
protobuf==3.18.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyotp==2.6.0
pyparsing==2.4.7
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sentry-sdk==1.4.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.7
Werkzeug==1.0.1
whitenoise==5.2.0
